add: custom exceptions and handler

+++ src/main/java/com/study/nclient_api/services/ClientService.java
@@ -2,14 +2,15 @@ package com.study.nclient_api.services;
 
 import com.study.nclient_api.entities.Client;
 import com.study.nclient_api.repositories.ClientRepository;
+import com.study.nclient_api.services.exceptions.DatabaseException;
+import com.study.nclient_api.services.exceptions.ResourceNotFoundException;
 import org.springframework.beans.BeanUtils;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.beans.Transient;
 
 @Service
 public class ClientService {
@@ -27,7 +28,8 @@ public class ClientService {
 
     @Transactional(readOnly = true)
     public Client getClientById(Long id) {
-        return clientRepository.findById(id).get();
+        return clientRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Recurso não encontrado"));
     }
 
     @Transactional
@@ -37,14 +39,25 @@ public class ClientService {
 
     @Transactional
     public Client updateClient(Long id, Client client) {
-        Client entity = clientRepository.getReferenceById(id);
-        BeanUtils.copyProperties(client, entity,"id");
-        clientRepository.save(client);
-        return entity;
+        try {
+            Client entity = clientRepository.getReferenceById(id);
+            BeanUtils.copyProperties(client, entity, "id");
+            clientRepository.save(client);
+            return entity;
+        } catch (Exception e) {
+            throw new ResourceNotFoundException("Recurso nao encontrado");
+        }
     }
 
     @Transactional(propagation = Propagation.SUPPORTS)
     public void deleteClientById(Long id) {
-        clientRepository.deleteById(id);
+        if (!clientRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+        try {
+            clientRepository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Falha de integridade referencial");
+        }
     }
 }
